Switch to unsigned 64-bit increment/decrement counters from Evan Miller and Dusgtin Sallings.
git-svn-id: http://code.sixapart.com/svn/memcached/trunk/server@619 b0b603af-a30f-0410-a34e-baf09ae79d0b

+++ thread.c
@@ -462,7 +462,7 @@ char *mt_defer_delete(item *item, time_t exptime) {
 /*
  * Does arithmetic on a numeric item value.
  */
-char *mt_add_delta(item *item, int incr, const unsigned int delta, char *buf) {
+char *mt_add_delta(item *item, int incr, const int64_t delta, char *buf) {
     char *ret;
 
     pthread_mutex_lock(&cache_lock);
